Move uri tests into own namespace
